refactor: enable passing of schema to the data classes for speed up
