+++ json_schema_generation_for_web_crawler/pypi_objs/dependency.py
@@ -1,7 +1,11 @@
 """
 Dependency Hyperedge Node
 """
-from .version import Constraint
+import re
+import typing
+from concurrent.futures import ThreadPoolExecutor
+from .version import Version, Constraint
+from .release import Release, Releases
 
 class Dependency:
     """
@@ -15,13 +19,14 @@ class Dependency:
         3.4. non-package extra
     """
 
-    def __init__(self, dep_constraint_str):
+    def __init__(self, dep_constraint_str, schema=None):
         """
         Input example:
         pytest (>=7.1.2,<8.0.0); (python_full_version > "3.6.0") and (extra == "test")
 
         ~=0.6 means >=0.6, ==0.*
         """
+        self._schema = schema
         if ';' in dep_constraint_str:
             dep_str, condition_str = dep_constraint_str.split(';')
         else:
@@ -32,7 +37,7 @@ class Dependency:
             ver_constraint_str = dep_str.split('(')[1].split(')')[0].strip()
             self._ver_constraint = Constraint(ver_constraint_str)
         else:
-            self._pkg = re.split(f'[<|>|=|!|~]', dep_str)[0].strip()
+            self._pkg = re.split(r'[<|>|=|!|~]', dep_str)[0].strip()
             self._ver_constraint = Constraint(dep_str.split(self._pkg)[1])
         self._condition_str = condition_str.strip()
 
@@ -40,23 +45,22 @@ class Dependency:
         return f'D[{self._pkg}@Ver[{self._ver_constraint}]@Cond[{self._condition_str}]]'
 
     @property
-    def depend_releases(self):
+    def depend_releases(self) -> typing.List[Release]:
         """
         Get sorted release data objs
         """
         def get_result(v):
-            return PypiVersionPackageView(self._pkg, v._ver_id)
+            return Release(self._pkg, v._ver_id, schema=self._schema)
 
-        with ThreadPoolExecutor(max_workers=4) as executor:
+        with ThreadPoolExecutor(max_workers=16) as executor:
             return list(executor.map(
                 get_result, self._depend_versions, chunksize=1))
 
     @property
-    def _depend_versions(self):
+    def _depend_versions(self) -> typing.List[Version]:
         versions = self._depend_package.versions
         return sorted(filter(self._ver_constraint.fit, versions))
 
     @property
-    def _depend_package(self):
-        return PypiPackageAllVersionView(self._pkg)
-
+    def _depend_package(self) -> Releases:
+        return Releases(self._pkg, schema=self._schema)
